+++ crates/hyperswitch_domain_models/src/router_request_types.rs
@@ -2,7 +2,7 @@ pub mod authentication;
 pub mod fraud_check;
 use api_models::payments::RequestSurchargeDetails;
 use common_utils::{
-    consts, errors, ext_traits::OptionExt, pii, types as common_types, types::MinorUnit,
+    consts, errors, ext_traits::OptionExt, pii, types as common_types, types::MinorUnit, id_type
 };
 use diesel_models::enums as storage_enums;
 use error_stack::ResultExt;
@@ -634,7 +634,7 @@ pub struct PayoutsData {
 
 #[derive(Debug, Default, Clone)]
 pub struct CustomerDetails {
-    pub customer_id: Option<String>,
+    pub customer_id: Option<id_type::CustomerId>,
     pub name: Option<Secret<String, masking::WithType>>,
     pub email: Option<pii::Email>,
     pub phone: Option<Secret<String, masking::WithType>>,