+++ crates/router/src/types/api.rs
@@ -366,7 +366,7 @@ impl ConnectorData {
                 enums::Connector::Opennode => Ok(Box::new(&connector::Opennode)),
                 // "payeezy" => Ok(Box::new(&connector::Payeezy)), As psync and rsync are not supported by this connector, it is added as template code for future usage
                 enums::Connector::Payme => Ok(Box::new(&connector::Payme)),
-                // enums::Connector::Payone => Ok(Box::new(&connector::Payone)), Added as template code for future usage
+                enums::Connector::Payone => Ok(Box::new(&connector::Payone)),
                 enums::Connector::Payu => Ok(Box::new(&connector::Payu)),
                 enums::Connector::Placetopay => Ok(Box::new(&connector::Placetopay)),
                 enums::Connector::Powertranz => Ok(Box::new(&connector::Powertranz)),
@@ -391,7 +391,7 @@ impl ConnectorData {
                 | enums::Connector::Plaid
                 | enums::Connector::Riskified
                 | enums::Connector::Threedsecureio
-                // | enums::Connector::Gpayments  Added as template code for future usage
+                | enums::Connector::Gpayments
                 | enums::Connector::Netcetera => {
                     Err(report!(errors::ConnectorError::InvalidConnectorName)
                         .attach_printable(format!("invalid connector name: {connector_name}")))